Move fullscreenImageWithCaption to widgets folder
