+++ lib/features/batches/view/batch_subpage.dart
@@ -82,7 +82,7 @@ class _BatchSubpageView extends StatelessWidget {
                       onTap: () {
                         Navigator.of(context).push(
                           MaterialPageRoute<dynamic>(
-                            builder: (_) => _FullscreenImageWithCaption(
+                            builder: (_) => FullscreenImageWithCaption(
                               heroTag: comm.name,
                               imageUrl: comm.img,
                               title: comm.name,
@@ -99,50 +99,6 @@ class _BatchSubpageView extends StatelessWidget {
           }),
         )
       ],
-    );
-  }
-}
-
-class _FullscreenImageWithCaption extends StatelessWidget {
-  const _FullscreenImageWithCaption({
-    required this.heroTag,
-    required this.imageUrl,
-    required this.caption,
-    required this.title,
-  });
-
-  final String heroTag;
-  final String imageUrl;
-  final String caption;
-  final String title;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: Navigator.of(context).pop,
-      child: Scaffold(
-        appBar: AppBar(leading: const SizedBox(), title: Text(title)),
-        body: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Hero(
-              tag: heroTag,
-              child: CachedNetworkImage(
-                imageUrl: imageUrl,
-                fit: BoxFit.contain,
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(8),
-              child: Text(
-                caption,
-                style: context.textTheme.labelMedium,
-                textAlign: TextAlign.center,
-              ),
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
