+++ lib/features/batches/view/widgets/widgets.dart
@@ -1,1 +1,2 @@
+export 'fullscreen_image_with_caption.dart';
 export 'image_with_caption.dart';
